+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.0 beta 3
 
+* The Hydra documentation URL is now `/apidoc` (was `/vocab`)
 * Automatic detection of the method of the entity class returning the identifier when using Doctrine (previously `getId()` was always used)
 * New extension point in `Dunglas\ApiBundle\Doctrine\Orm\DataProvider` allowing to customize Doctrine paginator and performance optimization when using typical queries
 * New `Dunglas\ApiBundle\JsonLd\Event\Events::CONTEXT_BUILDER` event allowing to modify the JSON-LD context